+++ source/DSCResources/SqlServer/SqlServer.schema.psm1
@@ -95,7 +95,7 @@ configuration SqlServer
     Import-DscResource -ModuleName SqlServerDsc -ModuleVersion 13.3.0
     . "$resourcePath\SqlServer.ScriptQuery.ps1"
 
-    Import-DscResource -ModuleName SecurityPolicyDsc -ModuleVersion 2.4.0.0
+    Import-DscResource -ModuleName SecurityPolicyDsc -ModuleVersion 2.10.0.0
     . "$resourcePath\Windows.SecurityOption.ps1"
 
     Import-DscResource -ModuleName AccessControlDsc -ModuleVersion 1.4.1