+++ source/DSCResources/WindowsClient/WindowsClient.schema.psm1
@@ -89,7 +89,7 @@ configuration WindowsClient
     Import-DscResource -ModuleName AuditPolicyDsc -ModuleVersion 1.4.0.0
     . "$resourcePath\windows.AuditPolicySubcategory.ps1"
 
-    Import-DscResource -ModuleName SecurityPolicyDsc -ModuleVersion 2.4.0.0
+    Import-DscResource -ModuleName SecurityPolicyDsc -ModuleVersion 2.10.0.0
     . "$resourcePath\windows.AccountPolicy.ps1"
     . "$resourcePath\windows.UserRightsAssignment.ps1"
     . "$resourcePath\windows.SecurityOption.ps1"